Add basic support for the Genbank data format (#12)
* start adding genbank data to the app

* make the log folder more cross-platform friendly

* cache deals only in streams so binary data can be stored

before, when trying to handle the streams early, things like the GZip file for the genbank file format data was becoming corrupted during storage/retrieval. The cache has been re-architected to deal only with streams, which allow the points of use the decide how to use the data. This allows us to successfully save + restore + decode a GZip file

* Simplify the project structure to allow for simpler builds

* Update assembly info to match old version

* add some data to the types

* update fsharp dependencies